Add files via upload
Python Script for "A frog is facing a line of 10 lily pads. With each jump, the frog moves forward, landing on one of the lily pads in front of it uniformly at random. What's the expected number of jumps required for the frog to reach the last lily pad?" from https://twitter.com/octonion/status/1630687525567070213